requirements.txt
roboflow
torch 
torchvision 
torchaudio
os
opencv-python
Pillow>=7.1.2
google-cloud-storage
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.2
gunicorn==20.1.0
Flask